+++ lib/view_models/field_cubits/image_field_cubit/image_field_cubit.dart
@@ -5,6 +5,7 @@ import 'package:cross_file/cross_file.dart';
 import 'package:equatable/equatable.dart';
 import 'package:flutter/foundation.dart';
 import 'package:i_validator/i_validator.dart';
+import 'package:reusable_editor/extensions/validation/x_file_image_validation_extension.dart';
 
 part 'image_field_state.dart';
 
@@ -12,7 +13,7 @@ class ImageFieldCubit extends Cubit<ImageFieldState> {
   ImageFieldCubit() : super(const ImageFieldState());
 
   void selectImage(XFile? pickedFile) {
-    final String? error = pickedFile?.path.validateImagePath();
+    final String? error = pickedFile?.validateImageFile();
 
     if (error == null) {
       emit(state.copyWith(pickedFile: pickedFile, imgFieldError: null));