+++ CHANGELOG.md
@@ -2,14 +2,17 @@
 
 All notable changes to this project will be documented in this file.
 
+
 ## 1.2.3
 
-### Aug 26, 2025
+### Sep 06, 2025
 
 ### ✨ Updated
 
 - Improved `AssetImageToFile` extension
 - Improved `AssetToXFileExtension` extension
+- Added XFile Image Validation Extension
+
 
 ## 1.2.2